rvl: match WPADWriteExtReg, BTA_Init, bta_sys_event, bta_dm_pm_timer_cback, AXFXDelayExpSettingsUpdateDpl2, GroupContainer dtor (us-80378840, us-802e09f0, us-802e1310, us-802e4814, us-802d9da0, us-8032ffb0) FULL_MATCH; bta_hh_add_device_to_list unit flags GC/3.0a5.2 + func_align 4; SDP_FindServiceUUIDInDb/OSPlayTimeRebootThread near-match

+++ libs/RVL_SDK/include/revolution/wpad/WPADInternal.h
@@ -273,7 +273,7 @@ BOOL WPADiSendGetContStat(WPADCommandQueue* pQueue, WPADInfo* pInfo,
                           WPADCallback pCallback);
 BOOL WPADiSendWriteDataCmd(WPADCommandQueue* pQueue, u8 cmd, u32 addr,
                            WPADCallback pCallback);
-BOOL WPADiSendWriteData(WPADCommandQueue* pQueue, const void* pSrc, u16 len,
+BOOL WPADiSendWriteData(WPADCommandQueue* pQueue, const void* pSrc, u32 len,
                         u32 addr, WPADCallback pCallback);
 BOOL WPADiSendReadData(WPADCommandQueue* pQueue, void* pDst, u16 len, u32 addr,
                        WPADCallback pCallback);
